Fixed various bugs in wrapper
MAINLY:
Image wrapper is now able to resolve event & room.
BUGS:
- image manager crashes whenever room / event isn't found.

+++ rec_net/managers/room/score.py
@@ -4,7 +4,7 @@ from attr import dataclass
 class Score:
     visit_type: int
     score: float
-    backup_score: int
+    backup_score: float
 
     @classmethod
     def from_data(cls, data):
